+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/TestEngine.kt
@@ -1,5 +1,6 @@
 package io.kotest.engine
 
+import io.kotest.core.config.Configuration
 import io.kotest.core.config.configuration
 import io.kotest.core.spec.Spec
 import io.kotest.engine.extensions.EmptyTestSuiteExtension
@@ -13,7 +14,9 @@ import kotlin.reflect.KClass
 data class TestEngineConfig(
    val listener: TestEngineListener,
    val extensions: List<EngineExtension>,
+   val configuration: Configuration,
 ) {
+
    companion object {
       fun default(): TestEngineConfig {
 
@@ -26,8 +29,13 @@ data class TestEngineConfig(
          return TestEngineConfig(
             listener = NoopTestEngineListener,
             extensions = engineExtensions,
+            configuration = configuration,
          )
       }
+   }
+
+   fun withConfig(configuration: Configuration): TestEngineConfig {
+      return TestEngineConfig(listener = listener, extensions = extensions, configuration = configuration)
    }
 }
 
@@ -41,6 +49,9 @@ data class EngineResult(val errors: List<Throwable>)
  */
 data class TestSuite(val specs: List<Spec>, val classes: List<KClass<out Spec>>)
 
+/**
+ * Multiplatform Kotest Test Engine.
+ */
 class TestEngine(val config: TestEngineConfig) {
 
    fun execute(suite: TestSuite) {